Improve Token structure: now only Value has a value property
This is a more proper way of representing a token -- Value tokens have a generic value property, and the others don't have a value property at all because they have a single representation. This way, if we've accessed a value property of a Token, we can be 100% certain that it is of type Value.
Now Token is not generic and, more importantly, the simple Tokens are no longer sub-types of 'Token<String>', which didn't make much sense.